# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorData:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorData'
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
  response.StandardResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorData'
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: bps1200@bps.go.id
    name: Tim Inovasi BPS Provinsi Sumatera Utara
    url: https://sumut.bps.go.id
  description: Backend RESTful API untuk Portal Aplikasi Terintegrasi BPS Provinsi
    Sumatera Utara.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sumut.bps.go.id
  title: Portal Aplikasi BPS Provinsi Sumatera Utara (ALUSI) API
  version: 1.2.0
paths:
...
  /admin/audit-logs:
    get:
      description: Mengambil log aktivitas operasional, perubahan data, dan akses
        administrator (Admin)
      parameters:
      - description: Filter ID pengguna/admin
        in: query
        name: user_id
        type: integer
      - description: Filter aksi (CREATE, UPDATE, DELETE, PROBE)
        in: query
        name: action
        type: string
      - description: Filter entitas target (app, category, guide, announcement)
        in: query
        name: entity
        type: string
      - description: Filter ID entitas spesifik
        in: query
        name: entity_id
        type: string
      - description: Filter tanggal awal (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Filter tanggal akhir (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: Halaman (default 1)
        in: query
        name: page
        type: integer
      - description: Jumlah item per halaman (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Jejak aktivitas admin (Audit Trail)
      tags:
      - Admin - Audit Trail
...
  /auth/callback:
    get:
      description: Menerima authorization code dari SSO BPS Sumut, menukar dengan
        token, dan melakukan JIT provisioning
      parameters:
      - description: Authorization code dari SSO
        in: query
        name: code
        required: true
        type: string
      - description: State token proteksi CSRF
        in: query
        name: state
        type: string
      - description: Set ke 'json' untuk mendapatkan JSON token langsung
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: OAuth2 / SSO Callback handler
      tags:
      - Auth
  /auth/login:
    get:
      description: Mengarahkan browser pengguna ke halaman login SSO BPS Sumut OIDC
        provider atau mengembalikan auth URL
      parameters:
      - description: Jika false, mengembalikan URL JSON tanpa redirect 302
        in: query
        name: redirect
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "302":
          description: Redirect to SSO Provider
          schema:
            type: string
      summary: Inisiasi login SSO BPS Sumut
      tags:
      - Auth
  /auth/logout:
    get:
      description: Menghapus cookie sesi lokal dan mengarahkan ke Single Log Out SSO
        BPS Sumut
      parameters:
      - description: URL tujuan setelah logout selesai
        in: query
        name: post_logout_redirect_uri
        type: string
      - description: Jika false, mengembalikan URL JSON tanpa redirect
        in: query
        name: redirect
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "302":
          description: Redirect to SSO Logout
          schema:
            type: string
      summary: Single Log Out (SLO)
      tags:
      - Auth
    post:
      description: Menghapus cookie sesi lokal dan mengarahkan ke Single Log Out SSO
        BPS Sumut
      parameters:
      - description: URL tujuan setelah logout selesai
        in: query
        name: post_logout_redirect_uri
        type: string
      - description: Jika false, mengembalikan URL JSON tanpa redirect
        in: query
        name: redirect
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "302":
          description: Redirect to SSO Logout
          schema:
            type: string
      summary: Single Log Out (SLO)
      tags:
      - Auth
  /auth/me:
    get:
      description: Mengembalikan informasi profil, role, dan izin pengguna yang sedang
        login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Profil pengguna saat ini
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan JWT token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"